ac1-example1-工序2： 实现KeepingEntity的申请场地功能

+++ src/main/java/com/txp/auctionbooking/application/dto/ApplyDTO.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ApplyDTO {
-    private Integer applyId;
+    private Long applyId;
     private String applyType;
     private String applyStatus;
     private String placeNumber;
